♻️ refactor: 앨범 모달 컴포넌트 gallery/components로 이동
