+++ src/pages/gallery/AlbumDetailPage/AlbumDetailPage.tsx
@@ -14,7 +14,7 @@ import {
 import { MENT_GALLERY } from 'constants/ments';
 import store from 'stores/RootStore';
 import Modal from 'components/common/Modal/Modal';
-import AlbumModal from 'components/common/AlbumModal/AlbumModal';
+import AlbumModal from 'pages/gallery/components/AlbumModal/AlbumModal';
 import useToast from 'hooks/common/useToast';
 import DataContext from '../context';
 import * as S from './AlbumDetailPage.styled';
