+++ src/pages/gallery/components/AlbumModal/AlbumModal.styled.ts
